% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{myTryCatch}
\alias{myTryCatch}
\title{Wrap an expr to suppress warnings and errors}
\usage{
myTryCatch(expr)
}
\arguments{
\item{expr}{}
}
\value{
a list object containing output, warnings and errors
}
\description{
Wrap an expr to suppress warnings and errors
}
\examples{
myTryCatch(log(-1))
}
