% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sssplugin-package.R
\docType{package}
\name{sssplugin}
\alias{sssplugin}
\title{RStudio Addins}
\description{
Use RStudio addins to execute \R functions interactively within RStudio.
See \code{\link{sssttest}()}
for an examples of R functions that are exported.
}
