% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetTest.R
\name{sssttest}
\alias{sssttest}
\title{Plugincode}
\usage{
sssttest(obj = NULL, ivar_name = NULL, dvar_name = NULL)
}
\arguments{
\item{obj}{the dataframe}

\item{ivar_name}{Indepedent variable}

\item{dvar_name}{depedent variable}
}
\value{
the t-test code
}
\description{
Plugincode
}
\examples{
df_test <- ToothGrowth
sssttest(df_test, "supp", "len")
}
