% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetTest.R
\name{sssttest}
\alias{sssttest}
\title{Interactive t-test}
\usage{
sssttest(obj = NULL, ivar_name = NULL, dvar_name = NULL)
}
\arguments{
\item{obj}{data frame to operate on}

\item{ivar_name}{name of the column for independent variable}

\item{dvar_name}{name of the column for dependent variable}
}
\value{
The function launches a shiny app in the system web browser. The recoding code is returned in the console
when the app is closed with the "Done" button.
}
\description{
This function launches a shiny app in a web browser in order to do
interactive t-test
}
\examples{
\dontrun{data(ToothGrowth)
sssttest(ToothGrowth)
}
}
\author{
André Calero Valdez <andrecalerovaldez@gmail.com>
}
